add teardown CLI option
