+++ smartsim/_core/utils/helpers.py
@@ -31,6 +31,7 @@ import base64
 import collections.abc
 import os
 import signal
+import subprocess
 import typing as t
 import uuid
 from datetime import datetime
@@ -286,6 +287,79 @@ def decode_cmd(encoded_cmd: str) -> t.List[str]:
     cleaned_cmd = decoded_cmd.decode("ascii").split("|")
 
     return cleaned_cmd
+
+
+def check_for_utility(util_name: str) -> str:
+    """Check for existence of the provided CLI utility.
+
+    :param util_name: CLI utility to locate
+    :returns: Full path to executable if found. Otherwise, empty string"""
+    utility = ""
+
+    try:
+        utility = expand_exe_path(util_name)
+    except FileNotFoundError:
+        print(f"{util_name} not available for Cray EX platform check.")
+
+    return utility
+
+
+def execute_platform_cmd(cmd: str) -> t.Tuple[str, int]:
+    """Execute the platform check command as a subprocess
+
+    :param cmd: the command to execute
+    :returns: True if platform is cray ex, False otherwise"""
+    process = subprocess.run(
+        cmd.split(),
+        capture_output=True,
+        check=False,
+    )
+    return process.stdout.decode("utf-8"), process.returncode
+
+
+def is_crayex_platform() -> bool:
+    """Returns True if the current platform is identified as Cray EX and
+    HSTA-aware dragon package can be installed, False otherwise.
+
+    :returns: True if current platform is Cray EX, False otherwise"""
+
+    # ldconfig -p | grep cray | grep pmi.so &&
+    # ldconfig -p | grep cray | grep pmi2.so &&
+    # fi_info | grep cxi\
+    ldconfig = check_for_utility("ldconfig")
+    fi_info = check_for_utility("fi_info")
+    if not all((ldconfig, fi_info)):
+        print("Unable to validate Cray EX platform. Installing standard version")
+        return False
+
+    locate_msg = "Unable to locate %s. Installing standard version"
+
+    ldconfig1 = f"{ldconfig} -p"
+    ldc_out1, _ = execute_platform_cmd(ldconfig1)
+    target = "pmi.so"
+    candidates = [x for x in ldc_out1.split("\n") if "cray" in x]
+    pmi1 = any(x for x in candidates if target in x)
+    if not pmi1:
+        print(locate_msg, target)
+        return False
+
+    ldconfig2 = f"{ldconfig} -p"
+    ldc_out2, _ = execute_platform_cmd(ldconfig2)
+    target = "pmi2.so"
+    candidates = [x for x in ldc_out2.split("\n") if "cray" in x]
+    pmi2 = any(x for x in candidates if target in x)
+    if not pmi2:
+        print(locate_msg, target)
+        return False
+
+    fi_info_out, _ = execute_platform_cmd(fi_info)
+    target = "cxi"
+    cxi = any(x for x in fi_info_out.split("\n") if target in x)
+    if not cxi:
+        print(locate_msg, target)
+        return False
+
+    return True
 
 
 @t.final
